+++ packages/react-ui/src/widgets/legend/LegendWidgetUI.js
@@ -1,5 +1,5 @@
 import React, { createRef, Fragment } from 'react';
-import { Box, Button, Collapse, Divider, Grid, SvgIcon, Typography } from '@mui/material';
+import { Box, Button, Collapse, Divider, Grid, SvgIcon } from '@mui/material';
 import makeStyles from '@mui/styles/makeStyles';
 import LegendWrapper from './LegendWrapper';
 import LegendCategories from './LegendCategories';
@@ -7,6 +7,7 @@ import LegendIcon from './LegendIcon';
 import LegendRamp from './LegendRamp';
 import LegendProportion from './LegendProportion';
 import PropTypes from 'prop-types';
+import Typography from '../../atoms/Typography';
 
 const LayersIcon = () => (
   <SvgIcon width='24' height='24' viewBox='0 0 24 24'>