+++ packages/react-ui/storybook/stories/organisms/InputFile.stories.js
@@ -1,15 +1,9 @@
-import {
-  Box,
-  Button,
-  CircularProgress,
-  FormControl,
-  InputLabel,
-  Typography
-} from '@mui/material';
+import { Box, Button, CircularProgress, FormControl, InputLabel } from '@mui/material';
 import React, { useState } from 'react';
 import InputFile from '../../../src/custom-components/InputFile';
 import { Alert } from '@mui/material';
 import UploadIcon from '../../../src/assets/UploadIcon';
+import Typography from '../../../src/atoms/Typography';
 
 const options = {
   title: 'Organisms/InputFile',