+++ packages/react-ui/storybook/stories/widgetsUI/WrapperWidgetUI.stories.js
@@ -1,10 +1,8 @@
 import React from 'react';
-import { Typography } from '@mui/material';
 import ColorizeIcon from '@mui/icons-material/Colorize';
 import MenuIcon from '@mui/icons-material/Menu';
 import AddLocationIcon from '@mui/icons-material/AddLocation';
 import WrapperWidgetUI from '.../../../src/widgets/WrapperWidgetUI';
-import { buildReactPropsAsString } from '../../utils';
 
 const options = {
   title: 'Custom Components/WrapperWidgetUI',