+++ packages/react-ui/src/widgets/ComparativeFormulaWidgetUI.js
@@ -37,14 +37,16 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 /**
- * Renders a <ComparativeFormulaWidgetUI /> widget
+ * Renders a `<ComparativeFormulaWidgetUI />` widget
+ * <!--
  * @param {Object} props
  * @param {number[]} props.data
  * @param {{ prefix?: string; suffix?: string; note?: string }[]} [props.labels]
  * @param {{ prefix?: string; suffix?: string; note?: string; value?: string }[]} [props.colors]
  * @param {boolean} [props.animated]
- * @param {{ duration?: number; animateOnMount?: boolean; }} [props.animationOptions]
+ * @param {{ duration?: number; animateOnMount?: boolean; initialValue?: number; }} [props.animationOptions]
  * @param {(v: number) => React.ReactNode} [props.formatter]
+ * -->
  */
 function ComparativeFormulaWidgetUI({
   data = EMPTY_ARRAY,