+++ packages/react-ui/storybook/stories/widgetsUI/ComparativeFormulaWidgetUI.stories.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import ComparativeFormulaWidgetUI from '../../../src/widgets/ComparativeFormulaWidgetUI';
+import { buildReactPropsAsString } from '../../utils'
 
 const options = {
   title: 'Custom Components/ComparativeFormulaWidgetUI',
@@ -9,9 +10,7 @@ const options = {
 export default options;
 
 const Template = (args) => <ComparativeFormulaWidgetUI {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {
+const sampleProps = {
   data: [1245, 3435.9],
   labels: [
     { prefix: '$', suffix: ' sales', note: 'label 1' },
@@ -19,3 +18,7 @@ Default.args = {
   ],
   colors: [{ note: '#ff9900' }, { note: '#6732a8' }]
 };
+
+export const Default = Template.bind({});
+Default.args = sampleProps;
+Default.parameters = buildReactPropsAsString(sampleProps, 'ComparativeFormulaWidgetUI');