+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- New Typography component to extend Mui Typography [#506](https://github.com/CartoDB/carto-react/pull/506)
 - New spacings, borders and shadows toolkit for MUI carto-theme [#507](https://github.com/CartoDB/carto-react/pull/507/)
 - Colors and Typography: design update [#502](https://github.com/CartoDB/carto-react/pull/502/)
 - Upgrade from Material UI v4 to MUI v5 [#501](https://github.com/CartoDB/carto-react/pull/501)